refactor: reuse dataframe code (part2)
