+++ src/main/scala/org/globalforestwatch/summarystats/carbonflux/CarbonFluxExport.scala
@@ -1,8 +1,7 @@
 package org.globalforestwatch.summarystats.carbonflux
 
-import org.apache.spark.sql.DataFrame
+import org.apache.spark.sql.{DataFrame, SparkSession}
 import org.globalforestwatch.summarystats.SummaryExport
-import org.globalforestwatch.summarystats.carbonflux.dataframes._
 import org.globalforestwatch.util.Util.getAnyMapValue
 
 object CarbonFluxExport extends SummaryExport {
@@ -14,24 +13,24 @@ object CarbonFluxExport extends SummaryExport {
     def exportWhitelist(df: DataFrame): Unit = {
 
       val adm2ApiDF = df
-        .transform(Adm2ApiDF.whitelist)
-        .coalesce(40) // TODO: optimize size so that tables have an avg file size of 100MB
+        .transform(CarbonFluxDF.whitelist(List("iso", "adm1", "adm2")))
+        .coalesce(1)
 
       adm2ApiDF.write
         .options(csvOptions)
         .csv(path = outputUrl + "/adm2/whitelist")
 
       val adm1ApiDF = adm2ApiDF
-        .transform(Adm1ApiDF.whitelist)
-        .coalesce(12) // TODO: optimize size so that tables have an avg file size of 100MB
+        .transform(CarbonFluxDF.whitelist2(List("iso", "adm1")))
+        .coalesce(1)
 
       adm1ApiDF.write
         .options(csvOptions)
         .csv(path = outputUrl + "/adm1/whitelist")
 
       val isoApiDF = adm1ApiDF
-        .transform(IsoApiDF.whitelist)
-        .coalesce(4) // TODO: optimize size so that tables have an avg file size of 100MB
+        .transform(CarbonFluxDF.whitelist2(List("iso")))
+        .coalesce(1)
 
       isoApiDF.write
         .options(csvOptions)
@@ -42,7 +41,7 @@ object CarbonFluxExport extends SummaryExport {
     def exportSummary(df: DataFrame): Unit = {
 
       val adm2ApiDF = df
-        .transform(Adm2ApiDF.sumArea)
+        .transform(CarbonFluxDF.aggSummary(List("iso", "adm1", "adm2")))
         .coalesce(40) // this should result in an avg file size of 100MB
 
       adm2ApiDF.write
@@ -50,7 +49,7 @@ object CarbonFluxExport extends SummaryExport {
         .csv(path = outputUrl + "/adm2/summary")
 
       val adm1ApiDF = adm2ApiDF
-        .transform(Adm1ApiDF.sumArea)
+        .transform(CarbonFluxDF.aggSummary2(List("iso", "adm1")))
         .coalesce(12) // this should result in an avg file size of 100MB
 
       adm1ApiDF.write
@@ -58,7 +57,7 @@ object CarbonFluxExport extends SummaryExport {
         .csv(path = outputUrl + "/adm1/summary")
 
       val isoApiDF = adm1ApiDF
-        .transform(IsoApiDF.sumArea)
+        .transform(CarbonFluxDF.aggSummary2(List("iso")))
         .coalesce(4) // this should result in an avg file size of 100MB
 
       isoApiDF.write
@@ -68,8 +67,12 @@ object CarbonFluxExport extends SummaryExport {
     }
 
     def exportChange(df: DataFrame): Unit = {
+      implicit val spark: SparkSession = df.sparkSession
+      import spark.implicits._
+
       val adm2ApiDF = df
-        .transform(Adm2ApiDF.sumChange)
+        .filter($"treecover_loss__year".isNotNull && $"treecover_loss__ha" > 0)
+        .transform(CarbonFluxDF.aggChange(List("iso", "adm1", "adm2")))
         .coalesce(100) // this should result in an avg file size of 100MB
 
       adm2ApiDF.write
@@ -77,7 +80,7 @@ object CarbonFluxExport extends SummaryExport {
         .csv(path = outputUrl + "/adm2/change")
 
       val adm1ApiDF = adm2ApiDF
-        .transform(Adm1ApiDF.sumChange)
+        .transform(CarbonFluxDF.aggChange(List("iso", "adm1")))
         .coalesce(30) // this should result in an avg file size of 100MB
 
       adm1ApiDF.write
@@ -85,7 +88,7 @@ object CarbonFluxExport extends SummaryExport {
         .csv(path = outputUrl + "/adm1/change")
 
       val isoApiDF = adm1ApiDF
-        .transform(IsoApiDF.sumChange)
+        .transform(CarbonFluxDF.aggChange(List("iso")))
         .coalesce(10) // this should result in an avg file size of 100MB
 
       isoApiDF.write
@@ -97,7 +100,7 @@ object CarbonFluxExport extends SummaryExport {
       getAnyMapValue[Boolean](kwargs, "changeOnly")
 
     val exportDF = df
-      .transform(ApiDF.unpackValues)
+      .transform(CarbonFluxDF.unpackValues)
 
     exportDF.cache()
